# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Tenwonmoa API
  description: Tenwonmoa API description
  version: 1.0.0
servers:
- url: http://3.39.184.232:8080
...
paths:
...
  /api/v1/account-book/search:
    get:
      tags:
      - api
      operationId: search-account-book
      parameters:
      - name: categories
        in: query
        description: 유저카테고리 아이디
        required: false
        schema:
          type: string
      - name: minprice
        in: query
        description: 최소 가격
        required: false
        schema:
          type: string
      - name: maxprice
        in: query
        description: 최대 가격
        required: false
        schema:
          type: string
      - name: start
        in: query
        description: 시작 등록일
        required: false
        schema:
          type: string
      - name: end
        in: query
        description: 종료 등록일
        required: false
        schema:
          type: string
      - name: content
        in: query
        description: "지출, 수입의 내용"
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: 페이지의 사이즈
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: 페이지 번호
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-account-book-search-1957235720'
              examples:
                search-account-book:
                  value: |-
                    {
                      "results" : [ {
                        "id" : 1,
                        "type" : "EXPENDITURE",
                        "amount" : 10000,
                        "content" : "점심",
                        "categoryName" : "식비",
                        "registerTime" : "2022-08-14T22:45:28.207416"
                      }, {
                        "id" : 1,
                        "type" : "INCOME",
                        "amount" : 50000,
                        "content" : "용돈",
                        "categoryName" : "용돈",
                        "registerTime" : "2022-08-14T22:45:28.207477"
                      } ],
                      "currentPage" : 1,
                      "totalElements" : 10,
                      "nextPage" : 2,
                      "totalPages" : 10
                    }
...
  /api/v1/incomes/:
    post:
      tags:
      - api
      operationId: income-create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-incomes-1988379506'
            examples:
              income-create:
                value: "{\"registerDate\":\"2022-08-14T22:45:25.767283\",\"amount\"\
                  :1000,\"content\":\"content\",\"userCategoryId\":1}"
              income-create-fail:
                value: "{\"registerDate\":\"2022-08-14T22:45:25.888186\",\"amount\"\
                  :1000,\"content\":\"content\",\"userCategoryId\":1}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-incomes--373420824'
              examples:
                income-create:
                  value: "{\"id\":1}"
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-expenditures-940438351'
              examples:
                income-create-fail:
                  value: "{\"messages\":[\"해당 사용자 카테고리는 존재하지 않습니다.\"],\"status\":404}"
  /api/v1/incomes/{incomeId}:
    get:
      tags:
      - api
      operationId: income-f
      parameters:
      - name: incomeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-expenditures-940438351'
              examples:
                income-find-by-id-fail:
                  value: "{\"messages\":[\"수입 정보가 존재하지 않습니다.\"],\"status\":404}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-incomes-incomeId-1674408626'
              examples:
                income-find-by-id:
                  value: "{\"id\":1,\"registerDate\":\"2022-08-14T22:45:25.938641\"\
                    ,\"amount\":1000,\"content\":\"content\",\"userCategoryId\":1,\"\
                    categoryName\":\"용돈\"}"
        "403":
          description: "403"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-expenditures-940438351'
              examples:
                income-forbidden:
                  value: "{\"messages\":[\"다른 사용자의 데이터에 접근할 수 없습니다.\"],\"status\"\
                    :403}"
    put:
      tags:
      - api
      operationId: income-update
      parameters:
      - name: incomeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-incomes-1988379506'
            examples:
              income-update:
                value: "{\"registerDate\":\"2022-08-14T22:45:23.10866\",\"amount\"\
                  :2000,\"content\":\"updateContent\",\"userCategoryId\":2}"
              income-update-fail:
                value: "{\"registerDate\":\"2022-08-14T22:45:25.649155\",\"amount\"\
                  :2000,\"content\":\"updateContent\",\"userCategoryId\":2}"
      responses:
        "204":
          description: "204"
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-expenditures-940438351'
              examples:
                income-update-fail:
                  value: "{\"messages\":[\"수입 정보가 존재하지 않습니다.\"],\"status\":404}"
    delete:
      tags:
      - api
      operationId: income-delete
      parameters:
      - name: incomeId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
...
  /api/v1/account-book/search/sum:
    get:
      tags:
      - api
      operationId: search-sum-account-book
      parameters:
      - name: categories
        in: query
        description: 유저카테고리 아이디
        required: false
        schema:
          type: string
      - name: minprice
        in: query
        description: 최소 가격
        required: false
        schema:
          type: string
      - name: maxprice
        in: query
        description: 최대 가격
        required: false
        schema:
          type: string
      - name: start
        in: query
        description: 시작 등록일
        required: false
        schema:
          type: string
      - name: end
        in: query
        description: 종료 등록일
        required: false
        schema:
          type: string
      - name: content
        in: query
        description: "지출, 수입의 내용"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-account-book-search-sum1751974973'
              examples:
                search-sum-account-book:
                  value: |-
                    {
                      "incomeSum" : 10000,
                      "expenditureSum" : 20000,
                      "totalSum" : -10000
                    }
...
components:
  schemas:
...
    api-v1-expenditures-940438351:
      type: object
      properties:
        messages:
          type: array
          description: 예외 메시지
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: number
          description: 에러 코드
...
    api-v1-account-book-search-1957235720:
      type: object
      properties:
        nextPage:
          type: number
          description: 다음 페이지
        totalPages:
          type: number
          description: 페이지
        currentPage:
          type: number
          description: 현재 페이지
        results:
          type: array
          description: "검색된 지출, 수입 데이터"
          items:
            type: object
            properties:
              amount:
                type: number
                description: 금액
              registerTime:
                type: string
                description: 등록일
              id:
                type: number
                description: 지출 or 수입의 아이디
              type:
                type: string
                description: 지출 or 수입 종류
              categoryName:
                type: string
                description: 카테고리 이름
              content:
                type: string
                description: 내용
        totalElements:
          type: number
          description: 전체 결과 개수
...
    api-v1-account-book-search-sum1751974973:
      type: object
      properties:
        totalSum:
          type: number
          description: "지출, 수입의 총합"
        incomeSum:
          type: number
          description: 수입의 총합
        expenditureSum:
          type: number
          description: 지출의 총합
...
    api-v1-incomes-1988379506:
      type: object
      properties:
        amount:
          type: number
          description: 수입 금액
        userCategoryId:
          type: number
          description: 유저 카테고리 ID
        content:
          type: string
          description: 내용
        registerDate:
          type: string
          description: 수입 등록 날짜
    api-v1-incomes--373420824:
      type: object
      properties:
        id:
          type: number
          description: 생성된 수입 아이디
...
    api-v1-incomes-incomeId-1674408626:
      type: object
      properties:
        amount:
          type: number
          description: 수입 금액
        userCategoryId:
          type: number
          description: 유저 카테고리 ID
        id:
          type: number
          description: 수입 ID
        categoryName:
          type: string
          description: 카테고리 이름
        content:
          type: string
          description: 내용
        registerDate:
          type: string
          description: 수입 등록 날짜
...